requirements.txt
flask
faiss-cpu
langchain
python-dotenv
langchain-core
langchain-openai
langchain-ollama
langchain-community
langchain-anthropic
langchain-google-genai
pandas
selenium
webdriver_manager
xlsxwriter
python-docx
requests
bs4
tqdm
os
pathlib